Add popup when entering an unknown word
When the user tries to submit an unknown word, they will be informed with a dialog box that the word is not in the dictionary.
If the user tries to submit words less than five letters long, they will not receive a pop-up as the unfilled columns should be enough information that their guess won't work.